+++ src/main/java/com/devbliss/gpullr/service/PullrequestService.java
@@ -10,10 +10,10 @@ import org.springframework.stereotype.Service;
 
 /**
  * Business facade for persisting and retrieving {@link Pullrequest} objects.
- * 
- * @author Henning Schütz <henning.schuetz@devbliss.com>
  *
+ * @author Henning Schütz <henning.schuetz@devbliss.com>
  */
+
 @Service
 public class PullrequestService {
 
@@ -29,14 +29,16 @@ public class PullrequestService {
 
   public List<Pullrequest> findAll() {
     return pullrequestRepository
-      .findAll()
-      .stream()
-      .sorted((p1, p2) -> p1.createdAt.compareTo(p2.createdAt))
-      .collect(Collectors.toList());
+        .findAll()
+        .stream()
+        .sorted((p1, p2) -> p1.createdAt.compareTo(p2.createdAt))
+        .collect(Collectors.toList());
   }
 
   public void insertOrUpdate(Pullrequest pullrequest) {
-    userRepository.save(pullrequest.owner);
+    if (userRepository.findOne(pullrequest.owner.id) == null) {
+      userRepository.save(pullrequest.owner);
+    }
     pullrequestRepository.save(pullrequest);
   }
 }